requirements-dev.txt
flake8==6.0.0
django-sslserver==0.22
factory_boy==3.2.1